Refactor Base/Title component to TypeScript
