+++ app/components/Base/Title/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Title';
